Refactor schema validation system: - Removed legacy validators and documentation for clarity. - Introduced explicit required and optional validators for better type safety. - Updated property definitions and normalization functions for label, name, and show props. - Enhanced controller methods to utilize new validation logic. - Consolidated README documentation for property definitions and usage examples.
