+++ packages/components/src/components/_skeleton/internal/functional-components/skeleton/controller.ts
@@ -1,8 +1,9 @@
 import type { LabelPropType } from '../../schema/props/label';
+import { normalizeLabel } from '../../schema/props/label';
 import type { NamePropType } from '../../schema/props/name';
-import { normalizeName, validateName } from '../../schema/props/name';
+import { normalizeName } from '../../schema/props/name';
 import type { ShowPropType } from '../../schema/props/show';
-import { normalizeShow, validateShow } from '../../schema/props/show';
+import { normalizeShow } from '../../schema/props/show';
 import { BaseController } from '../base-controller';
 import { ClickButtonController } from '../click-button/controller';
 import type { ControllerInterface, WebComponentInterface } from '../generic-types';
@@ -24,21 +25,22 @@ export class SkeletonController<
 	}
 
 	public watchLabel(value?: LabelPropType): void {
-		this.clickButtonController.watchLabel(value);
+		// Use normalize function for label
+		const processedValue = normalizeLabel(value);
+		this.setRenderPropsOrStates('label', processedValue);
+		this.clickButtonController.watchLabel(processedValue);
 	}
 
 	public watchName(value?: NamePropType): void {
-		const normalized = normalizeName(value);
-		if (validateName(normalized)) {
-			this.setRenderPropsOrStates('name', normalized);
-		}
+		// Use normalize function for name
+		const processedValue = normalizeName(value);
+		this.setRenderPropsOrStates('name', processedValue);
 	}
 
 	public watchShow(value?: ShowPropType): void {
-		const normalized = normalizeShow(value);
-		if (validateShow(normalized)) {
-			this.setRenderPropsOrStates('show', normalized);
-		}
+		// Use normalize function for show
+		const processedValue = normalizeShow(value);
+		this.setRenderPropsOrStates('show', processedValue);
 	}
 
 	public toggle(): void {
